Implementado Service.
