+++ StockServiceWCF/App_Code/IStockService.cs
