+++ StockServiceWCF/App_Code/Products.cs
