--- StockServiceWCF/App_Code/Repositories/StockService.cs
@@ -1,42 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using MetalBake.Models;
-namespace MetalBake
-{
-    public class StockService : IStockService
-    {
-        public Dictionary<string, int> productsStock = new Dictionary<string, int>()
-        {
-            {"B",40 },
-            {"M",36 },
-            {"C",24 },
-            {"W",3 }
-
-        };
-        Products products = new Products();
-        public  void CheckStock(string[] itemId)
-        {
-            foreach (var item in itemId)
-            {
-                if (productsStock[item] <= 0)
-                {
-                    Console.WriteLine($"Not enough stock of {products.productsNames[item]}");
-                    break;//Uso exit para parar aplicacion
-                }
-                else
-                {
-                    productsStock[item]--;
-                }
-            }
-
-        }
-        public void ShowStock(string[] itemId)
-        {
-            foreach (var item in itemId)
-            {
-                Console.WriteLine($@"{products.productsNames[item]}  ||  {productsStock[item]}");
-            }
-        }
-    }
-}
